fix initialization of session state
