+++ scripts/constants/constants.py
@@ -23,6 +23,6 @@ class SelectionStatus(Enum):
 NOTEBOOK_UPLOAD_STATUS = {
 
     WidgetText.WARNING.value:'存在当前未翻译的 Notebook',
-    WidgetText.ERROR.value:'无法解析上传的 Notebook'
+    WidgetText.ERROR.value:'无法解析上传的 Notebook',
 
 }
